Update patch set 2
Patch Set 2:

I ran a search on the LineageOS github - the ONLY devices that contain the fileencryption flag in fstab are this one and marlin (Pixel). The rest, almost a hundred devices running Android 7.1 which supports FBE, do not have it enabled in their device configuration. This is obviously a mistake that was caused by copying and pasting the fstab file from OxygenOS.

The bigger issue at hand is that there might not be a way to undo this. Because of the intrusive nature of FBE, it cannot be enabled optionally, and if we just disable it all current users will end up with a corrupt userdata. Does mounting a FBE-encrypted partition without the fileencryption flag fail the mount? If so, perhaps we can hack it with two mount lines in fstab, similar to what we used to do to support both ext4 and f2fs?

Patch-set: 2
Reviewer: Yuval Deutscher <17855@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0